cmd: updated version handling
- Added the version command to replace the --version flag
  for the check command;

- Added a new field for version, now the version command
  will also show the NoVerify version number;

- Build time will now be shown with delimiters.
